+++ client/src/containers/analysis-chart/impact-chart/component.tsx
@@ -153,7 +153,7 @@ const StackedAreaChart: React.FC<StackedAreaChartProps> = ({ indicator }) => {
   }, []);
 
   return (
-    <div className="p-6 bg-white rounded-md shadow-sm">
+    <div className="p-6 bg-white rounded-md shadow-sm" data-testid="analysis-chart">
       {isFetching && <Loading className="w-5 h-5 m-auto text-navy-400" />}
       {!isFetching && isFetched && data && (
         <div>