+++ client/src/hooks/impact/comparison.ts
@@ -28,8 +28,6 @@ export type ImpactComparisonParams = {
   scenarioId?: Scenario['id']; // TO-DO: remove and unify with comparedScenarioId
   baseScenarioId?: Scenario['id'];
   comparedScenarioId: Scenario['id'];
-  scenarioOneId?: Scenario['id']; // TO-DO: remove and unify with baseScenarioId
-  scenarioTwoId?: Scenario['id']; // TO-DO: remove and unify with comparedScenarioId
   indicatorIds: string[];
   groupBy: string;
   startYear: number;
@@ -87,7 +85,8 @@ export const useImpactScenarioComparison = <T = ImpactDataApiResponse<'scenario'
     ['impact-scenario-vs-scenario', typeof params]
   > = {},
 ) => {
-  const enabled = (options.enabled ?? true) && !!params.scenarioOneId && !!params.scenarioTwoId;
+  const enabled =
+    (options.enabled ?? true) && !!params.baseScenarioId && !!params.comparedScenarioId;
 
   const query = useQuery(
     ['impact-scenario-comparison', params],